Разбил по файлам, добавил загрузку фильмов по кнопке
